{"types":"C","runs":72}

+++ Problem15.java
@@ -5,8 +5,8 @@
 import acm.program.ConsoleProgram;
 public class Problem15 extends ConsoleProgram{
 	public void run() {
-		double a = readInt("Enter a: ");
-		double b = readInt("Enter b: ");
+		int a = readInt("Enter a: ");
+		int b = readInt("Enter b: ");
 		double average = (a + b) / 2;
 		println("The arithmetic average is " + average + ".");
 	}
